+++ app/src/main/java/io/github/muntashirakon/AppManager/backup/dialog/BackupRestoreDialogFragment.java
@@ -399,7 +399,8 @@ public class BackupRestoreDialogFragment extends CapsuleBottomSheetDialogFragmen
         ContextCompat.registerReceiver(mActivity, mBatchOpsBroadCastReceiver,
                 new IntentFilter(BatchOpsService.ACTION_BATCH_OPS_COMPLETED), ContextCompat.RECEIVER_NOT_EXPORTED);
         // Start batch ops service
-        BatchBackupOptions options = new BatchBackupOptions(operationInfo.flags, operationInfo.backupNames, operationInfo.relativeDirs);
+        BatchBackupOptions options = new BatchBackupOptions(operationInfo.flags, operationInfo.backupNames,
+                operationInfo.relativeDirs, operationInfo.exclusionGlobs);
         BatchQueueItem queueItem = BatchQueueItem.getBatchOpQueue(operationInfo.op,
                 operationInfo.packageList, operationInfo.userIdListMappedToPackageList, options);
         Intent intent = BatchOpsService.getServiceIntent(mActivity, queueItem);